9/11

+++ src/main/java/com/example/pbl_api/service/IBillService.java
@@ -9,7 +9,7 @@ public interface IBillService {
 
     List<BillModel> getAllBills();
 
-    List<BillModel> getBillsByUser(int idUser);
+    List<BillModel> getBillsByUser(long idUser);
 
     BillModel saveBill(List<ProductModel> productList,int idUser,double total,boolean type,List<Integer> productsAmountList);
